Reorganize Trainer Flags & More Rosters
This is a save break.

Trainer flags have been reorganized. The order is Johto -> Kanto -> rematches -> flags that could potentially be purged. The Johto flags are in roughly the order of an area's appearance, while the Kanto flags are routes first then gyms. Most of the Kanto flags and all the rematch flags are still placeholders. The `opponents.h` file has comments separating the blocks of flags by area.

Filled out the rest of the Rival rosters, trainers in all of Kanto's Gyms (including the Leaders), Cal, and Red. `set_gym_trainers.inc` has been completely filled out with all gym trainers.

Also nuked the rematch struct, as it was annoying and has to be totally redone anyway. Will fill it back out when rematch rosters are done.