Do All Rematch Rosters
All trainer rosters from Crystal are now in.

Trainers involved in rematches have had their constants and labels renamed to include the number suffix.

The rematch roster defines are commented with the event that gates that roster, for later reference.

We're at 480 trainers now, which is 60 bytes exactly. The 89 trainer flags at the end could potentially be saved to reduce this further.