Script Most Routes Pre-Champion
This will not compile until my other PR is merged, as it uses the new opponent constants.

Maps or events that require flags to be set aside have not been scripted. This includes MooMoo Farm, the Suicune event on Route 42, the Route 43 Gatehouses, Red Gyarados and the Lance Event, the houses in the Lake of Rage, the Dragon Shrine, Kiyo in Mt. Mortar, the Sandstorm TM house on Route 27, and the Battle Tower gatehouse. When scripts have been omitted on an otherwise scripted map, there is a TODO detailing what is missing.

No item balls have been scripted either.

Routes 38, 39, 40, 41, 42, 43, 44, 45, 26, and 27 have been scripted. The Lake of Rage, Mt. Mortar 1F and 2F, and Dragon's Den have also been scripted. Route 37 has not been touched, as last I heard ExpoSeed was working on that map.

All trainers that have phone calls have a TODO comment in their scripts.