Script Most of Olivine City & Lighthouse
Again, things that need flags are not scripted. This includes the Rival event, Jasmine and her Gym Guy in the Gym, the in-game trade house, the top-floor of the Lighthouse, and the ferryman at the port. They are marked as TODO in the relevant files.

The Strength HM and Good Rod are obtainable. The scripts are set up so no flags are needed.

Again, no item balls are done.