Script Most of Blackthorn City & Indigo Plateau
In Blackthorn, Clair, the Gym boulders, the Gym and Dragon's Den guards, and the in-game trade house are not scripted.

The Victory Road badge check is not scripted.

In Indigo Plateau, the Rival rematch isn't scripted.

Additionally, areas with marts have had their marts filled out, with the exception of Cianwood and Mahogany. Olivine's mart sells Surf Mail in Crystal, but sells no mail in CD for now.