Map Routes 26-28, Tohjo Falls, & Mt. Silver's Exterior
Still need to randomize the grass, but these routes are now finished. The warps to Tohjo Falls are in place as well as the warp to Mt. Silver's PMC but no other events are.

Ended up changing the color of the doors in New Bark Town because I ran out of palettes on that tileset. The regular houses now have green doors instead of red.