Fix Tile Issues with Route 26
The player should now appear above the bridges on Routes 26 and 27, one of the ledges can now not be walked through, a ledge up north of Route 26 will no longer jump the player into the mountain, the player is now drawn above the pillars at the entrance to Victory Road.

The Victory Road entrance gate building is totally illuminated at night instead of the windows, so that needs to be fixed.