Do Met Locations
Because Colosseum and XD don't accept Pokemon from invalid met games, met location handling had to be redone. All Pokemon from CD are set to have Emerald met games now.

CD-exclusive locations are 0xD5 and up, with some entries doubling up, as there are more than 0xFF met locations with all the new Johto ones. Pokemon from these CD-exclusive locations will show "obtained in a trade," in the vanilla games.

Entries are doubled up by checking/setting a bit in the filler field of the Pokemon data structure. This "new" field is called locationBit and the param for the get and setmondata function is MON_DATA_LOCATION_BIT.

Map sections have not been reshuffled, instead a lookup table sets the appropriate met location to generated Pokemon/Eggs and if the met location is one of the doubled-up ones, it sets the location bit appropriately.

The summary screen handles showing locations from all vanilla GBA games and the CD-exclusive locations, and the special Egg locations for the Togepi and Odd Eggs are handled.