Map Route 43 & Lake of Rage
Route 43 and the Lake of Rage are mapped. The houses and gatehouses have their warps set up as well. Lake of Rage has trainer tips signs to mark where the cut trees should be, as the headbutt tree maze has been slightly altered.

As always, grass still needs to be randomized. Route 43 may receive some additional tweaks in the future as well as I'm not totally happy with it.