Totally Finish Route 37
Route 37 is mapped, evented, and scripted. It is 100% playable. Grass needs to be randomized.

Removed Sierra's mini-credits event and added in the missing Trainer Tips sign on Route 36.

The southern entrance to Ecruteak has been modified slightly to facilitate connecting the maps., as was the northern exit of Route 36.

Additionally, Violet City's tileset was modified to prevent some issues when transitioning between Routes 36 and 37 which resulted in some tiles being shuffled around. The Ruins of Alph outside map was modified because of this, but is visually identical.