Implement Roamers
Raikou and Entei are now released properly during the Burned Tower event. They roam and are assigned the Legendary Beast battle theme left over from RSE as a placeholder.

Mechanically, the roamers work like they do in Gen 3 with the exception that they will not appear in water, only in tall grass.