Fix Tile Errors & Missing Tileset Fields
Tile errors have been fixed on the maps for Blackthorn City, Lake of Rage, and Routes 41, 42, and 44.

Route 42's still pond water has been replaced with the ocean water because of issues with the reflections. Lake of Rage needs to be edited for a similar reason, as the top of the headbutt trees aren't correct when surfing above them.

Mahogany and Blackthorn tilesets were missing their tileset overrides in their headers, so this was fixed which eliminated lag on maps that use those tilesets.