Fix Link Restrictions
Initially, CrystalDust can only trade with itself. The Bill event in Ecruteak's Pokemon Center queues the Time Capsule to be unlocked on the next day. Once that happens, link restrictions (including with Colo, XD, and Box) are completely dropped.

Additionally, not having the National Dex should now never be an impediment to trading as long as the Time Capsule restrictions are followed.

The error message has been changed to mention not being ready to trade across time yet if the Time Capsule isn't unlocked.

Similar to Gen 3, battles are always allowed regardless of the current trade restrictions.

Existing saves already have the Gamecube linking enabled due to Sierra mistakenly enabling it when getting the Pokedex.