Do Most of Olivine City & Route 40
Olivine City is scripted and has its objects placed save for the Lighthouse, Gym, and the ferryman in the port. The door animations still need to be inserted.

Olivine Port and Cafe have been mapped and are working fully. Olivine Gym is only partially mapped, as it needs a new tileset.

Route 40 has had its objects placed and scripts hooked up. The Gatehouse currently has an officer with no script blocking off the north exit until we decide what's happening with the Battle Tower/Frontier.

Additionally, filled out the in-game trades because I had to for the trade Voltorb here.

Also fixed the HM gates based on Badges. Rock Smash no longer needs a Badge to use. Strength needs the Plain Badge. The Fog Badge allows the player to Surf in the field without selecting it from the menu now.

Ecruteak City now sets its visited flag when entered; it was missed when scripting the city.