Script Most of Cianwood City
The only thing that hasn't been scripted is the Suicune encounter.

The Poke Seer is the elemental Hyper Beam tutor. The Photo Studio is set up to put the player's party on the back of their Trainer Card and will take the player's money to do so, but actually doing it is not yet implemented.

The Secret Potion was created to finish the pharmacy, and new specials were implemented for the Poke Seer tutor. Tutor learnsets had to be reformatted because there is now more than 32, but other than the addition of the elemental Hyper Beams to the Johto starters, the learnsets themselves have been left alone.

Also fixed a bug introduced while implementing the new fast jump movements. Jump movements should now all work correctly.