Backfill Union Cave & Slowpoke Well
Union Cave B2F and Slowpoke Well B1F are done. Lapras can be encountered and only appears on Friday.

Additionally, fixed a bug with the Eusine palette override for Burned Tower 1F that messed with the player's palette and fixed a typo with Calvin's intro text in Union Cave B1F.