Do OW Suicune Encounters & Sudowoodo Fleeing
All the Suicune encounters in the overworld are done, including the Eusine battle. While doing this, made the events in the Burned Tower more accurate to Crystal. The Rival disappears after his defeat, and his defeat triggers the hole to stay. Morty only leaves when the beasts are awakened and Eusine will stay in B1F if not talked to (using Escape Rope after awakening the beasts, for example) until going to Cianwood like in Crystal.

Sudowoodo now flees into the trees if the player runs or teleports.

Additionally, flags have been reserved for regenerating uncaught on-offs. Sudowoodo's script now includes that handling, as well as Raikou and Entei when their roamer battles are over.

There is an issue with field effect palettes overwriting one another. This is most evident with Cianwood's Suicune encounter with the footprints, exclamation point, and surf blob palettes all recently loaded. Will be fixed later.