Do Mahogany Rocket Hideout
The Mahogany Hideout is scripted and finished in its entirety. The final Executive battle is a multi battle with Lance like in HGSS, and the Executive's partner has the same roster as in HGSS. Lance's Dragonite has been manipulated to appear to be the Extremespeed gift Dratini from the Dragon's Den, as a little Easter egg.

Because the trap encounters in the hideout needed to  prevent running from battle, that functionality was implemented. The Red Gyarados now correctly cannot be run from.

Due to the large amount of flags needed for the hideout, some unused Emerald scripts were cleared out, as they used replaced flags.