Map Champion Room & Modify Koga's Room
Lance's room has been mapped. It needed a whole new tileset because there were not enough palettes  left on the PokemonLeague tileset.

Because Lance's room was put on its own tileset, the PokemonLeague tileset removed all of those tiles and was reduced. This also freed up a palette, giving enough room to put tall grass tiles in for Koga's room.