Script Radio Tower
The entirety of the Rocket Takeover event that occurs in the Radio Tower is scripted, as well as the post-takeover events in the tower.

The part of the takeover that occurs in the Goldenrod Underground is not scripted, and the door in the Underground can still not be opened despite the Basement Key now being obtainable.

Also inserted a Meowth OW sprite because I forgot that I needed one for Mary's radio booth when I inserted all the OWs.