Map Goldenrod Underground for Takeover
The Underground's switch room and warehouse are mapped. The switch room's shutters have not been set up in any way, but the metatiles for it are present in the tileset.

Made a minor adjustment to Goldenrod Dept. Store B1F to make the warp to the warehouse consistent and to remove a doodad that was purged to make more room for Radio Tower doodads, as these maps use the same tileset.