Script Clear Bell Event to Battling Suicune
The Tin Tower Gate, Wise Trio's Room, And Tin Tower 1F are now fully scripted. Suicune plays the VS Legendary Beasts theme for its battle. The Wise Trio's room has the two trainers teaming up to battle functionality disabled so it flows better. The Trio must be fought one at a time. The post-HoF unlocking of the rest of the Tin Tower and receiving the Rainbow Wing when the player has all the legendary beasts is implemented but not tested.

Their is a strange, split-second stuttering of the music when exiting Suicune's battle due to using turnobject in an ON_RESUME map script. Where that happens is commented in the poryscript.

Additionally, the Tin Tower Gate and Wise Trio's Room pre-Clear Bell were not scripted, so that was taken care of in this commit.