Implement OW Whirlpools & Whirlpool Field Move
Whirlpools have been placed in the OW. They animate, and can be dispersed by using Whirlpool as a field move.

In party_menu.c, there is now a function that checks badges for HM use explicitly rather than relying on the order of the field move constants. Rock Smash now no longer requires a badge to use from the party as a result, like in GSC.