Script Up to Elite Four & Fix Met Location Bug
Everything up to the Elite Four should now be scripted, including Kiyo in Mt. Mortar and the Rival Rematch in Indigo Plateau.

Certain met locations were not being written properly to Pokémon. I'm pretty sure that this fixes all the cases where that occurs.

Tweaked some text on Routes 26 and 27, and the Abra man in Indigo Plateau.

Also fixed an issue with getting Dratini in the Dragon Shrine.