Script Elite Four Up to Hall of Fame
The Elite Four, Champion, and Hall of Fame sequence is now scripted. Elm will call and give the S.S. Ticket once the player reloads back into New Bark.

One offs are properly regenerated, including roamers, if not caught.

Also fixed a bug with the function that checks if the player owns all the beasts with their OT.