(Mostly) Finish Tin Tower
Tin Tower is complete, save for the placeholder map for the roof. Ho-oh can be caught if the player has become the Champion and has all three legendary beasts with their OT in either the party or boxes. Like in Crystal, the game doesn't check the Day Care for them.