Do Most of Vermilion City
Vermilion City is mostly scripted and its events are set up.

Snorlax is not scripted, and it is placed in such a way that it doesn't actually block anything.

Vermilion's map has been altered to move Diglett's Cave to it, but the port building still needs to be added to the tileset and map. Route 11's map has been altered as well to smoothly connect to the new Vermilion.

Vermilion also needs a Clefairy Doll, Machop, and Bayleef OW to be complete. These have not been inserted yet.

Door and tile animations do not yet play.

All Kanto towns have their map flags set in sMapSecFlags.

Saving and loading the game will now count Kanto badges as well as Johto badges for displaying the badge count.