Enable Kanto Map; Fix Fly
The region map now switches based on region.

Using Fly in the field now brings the player to the correct spots.

Some Kanto locations have had their places assigned to the Kanto region map, but the region map needs to be edited to account for FRLG map sizes so I stopped midway.