Map Mt. Moon
The GSC Mt. Moon and Mt. Moon Square is mapped. The Mt. Moon gift shop is using the Mahogany shop layout and will need a setmetatile script to always hide the stairs.

FRLG's Mt. Moon maps have been imported and labeled as B1F, B2F, and B3F with GSC's Mt. Moon as 1F.

All warps have been placed and connected, but with the exception of the shop and the sign outside of it no other events have been placed.