Fix Errors & Add Missing Door Anims
Fixed some typos in recent texts in Kanto.

Fixed tile error on the outdoor Seafoam Islands.

Added door animations for the Moomoo Farm and Barn, Olivine, Cianwood, Lighthouse, and Goldenrod Underground.

Hatched Eggs now do not overwrite the met game as it is no longer necessary with how met locations are done. This is vanilla behavior.