Do Most of Cerulean City
All of Cerulean is done except for the Gym and for inserting Zubat and Diglett OWs. Both OWs have images in the folder, but they are not built or hooked up to anything.

Also alphabetized the Pokemon in spritesheet_rules.mk because why not?

Committing now because I need to expand OWs and want to do that in one commit.