Do Silver Cave, Fix Some Tile Errors, Fix E4 & Champ mugshots
Silver Cave is fully done. Red can be battled. His flag was repurposed from the unneeded hide flag that the Magnet Train president in Goldenrod had, as it is already set at the start of the game. The president is hidden correctly, and Red's hide flag is now unset by entering the Hall of Fame.

Fixed collision errors on Route 19 and tile errors and reflection issues on Route 28 and Silver Cave's exterior.

The Elite Four and Champion now show the correct trainer pics in their mugshot transitions.

The Champion now uses the FRLG Champion theme (as does Red).