Do FRLG Mt. Moon
The lower three floors of Mt. Moon are fully done. This consists of the maps ported from FRLG. Most of the trainers were taken from HGSS's Route 3, with two of the Firebreathers being taken from GS's Burned Tower. Two of the trainers are custom to fill out the map a little more.

There are hidden mushrooms where they were in FRLG that will regenerate everyday.

The wild encounters are beefed up to be below the levels of Silver Cave's exterior as this is technically an optional area, making it a good place to train before opening Mt. Silver.

The trainers' levels are around (but not above) the Mt. Moon Rival's levels.