Do Goldenrod Game Corner Move Tutor
The post-Champion move tutor outside the Goldenrod Game Corner is now implemented. Because all the moves he taught in Crystal are Gen 3 TMs, he sells a TM for coins instead of directly tutoring a move. His text had to be edited slightly to accommodate for this change.

Sierra reserved a flag to hide him that is unnecessary, so I converted it to an unused flag. It is still set in `new_game.inc` and old players will have it set in their save files, so it can be used for something else that should've been set from the very beginning.