Insert SS Aqua OW, Fix DN Tint in SS Aqua, Fix SS Aqua on Town Map
The SS Aqua now has an OW. It looks misplaced in porymap but appears correctly in-game.

Inside the SS Aqua was incorrectly tinting the palettes based on the time. These areas now use the regular building tileset for their primary tileset to fix this.

The town map now displays on the correct coordinates when inside the SS Aqua, but the special icon for it is still not implemented.

Additionally, the captain's room now has maps of both regions on the walls for the aesthetics. Only the Johto one is interactable.