In-Game Trade Machop Holds Macho Brace
In HGSS, Muscle the in-game trade Machop obtainable in the Goldenrod Department Store is holding the Macho Brace. This is as good a place as any to place this item, so it has been done. This Machop originally held Gold/Sitrus Berry, which is renewable so there is no loss here.