Create New String Placeholders, Save ~14.7KB in ROM With Them
String placeholders have been created for the "POKéMON", "POKéDEX", and "POKé" strings. Every instance of these strings in script-related files has been replaced. This was 2841 occurrences for POKéMON, 45 for POKéDEX, and 272 for POKé. This saves about 14.7KB in the ROM altogether, but more importantly makes it so I never have to find and copy+paste the accented é character ever again.

These new string placeholders replace the ones for Archie, Maxie, and Kyogre.