Handle Fainting to Rocket Hideout Electrode
Like in Crystal, walking into the generator room without having beaten all the Electrode will force the player to walk to the left side.

Due to the bigger screen size, Lance is present on the right side if viewed from other parts of the Hideout like in HGSS. This was not an issue in Crystal, because most of the right side of the generator room could not be seen.