Fix Route 14 Cut Tree Not Spawning Correctly
When coming from Route 13, the northern cut tree on Route 14 wasn't spawning correctly. This is due to the in_connection code being a little too zealous when setting temp flags. Since it only affected this one tree, I opted to assign it a temp flag out of the existing code's range so it isn't set incorrectly when transitioning from Route 13 to Route 14.