Implement SS Aqua Region Map Icon
The Fast Ship now has an icon that appears on the region map instead of the player's head when the player is onboard, like in Crystal. The icon also faces the direction the ship is sailing.

While on the ship, the region map cursor is now initialized on New Bark Town like in Crystal.