Fast Ship Captain's Room Aesthetic Changes
The new Fast Ship icon doesn't look good when viewing the region map outside of the PokeGear (like through a map metatile on the wall). The Fast Ship captain's room had a map on the wall for decoration that was interactable. This map has been replaced with a porthole and now the non-interactable map will swap between the Kanto and Johto region maps depending on the current direction of travel.