Fix Route 42 & 43, Blackthorn Tile Errors, Change Blackthorn's Music
Route 42 had a tree shadowing error that was fixed.

Route 43's gatehouse's left corner of the roof had a green pixel in the middle of the yellow path which was fixed.

The Move Deleter's house had incorrect shading on the rocks behind it due to missing triple-layer tiles. Sacrificed an unused palette to fake the effect instead and fix it.

Additionally, Blackthorn City now uses the FRLG version of its theme so that it and Azalea Town have distinct music.