Dragon's Den Shrine Tweaks + Fix Night Tinting For Some Maps
Dragon's Den has had aesthetic tweaks to its tileset. The Gyarados statues are now generic dragons. There are animated torch fires and light puddles. Also changed the Elder back to the other kind of old man. The shrine and the cave entrance are now tinted for night as well.

Additionally, discovered that all of the places that I assigned perpetual night to weren't working correctly, which has been fixed.