Room Decoration Work
Fixed the palettes of all object event decorations to match FRLG's palettes. Also fixed UltraLight's messed up palettes for the dolls that they contributed.

Added the Gold and Silver Shield object events and special handling for their palettes in the room.

There is now space in gSaveBlock1 reserved for the room decor inventory. The save has not been shifted.

Default decor is set to the room and put in the decor inventory when a new game is started.

The game now handles decorating the room based on the contents of the VAR_ROOM_* variables. The player's room metatile set was altered to make this easier to code.

All decorations that should have interact text do. New metatile behaviors were created to facilitate this for the posters and consoles.

There is still no way to obtain decorations (other than the ones obtained on new game) and no menu to actually decorate with.