Fix Saffron/Route 7 & Route 6 Connection Inconsistencies
The vertical fence under the gatehouse on the Saffron side now ends under the treetop to match Route 7's side.

The trees on both sides of the Route 6 gatehouse had incorrect tree shadowing, which created an inconsistency with the Saffron side which was fixed.

These connections are illusions, so both sides need to be matched tile-wise.