Do Unlocking Frontier & Frontier Port
The player is now invited to the Battle Frontier by Scott on the SS Aqua before they enter Vermilion for the first time.

Once that happens, the gatehouse to what *was* Crystal's Battle Tower opens and leads to a port. From there, the player can sail to the Battle Frontier.

The Battle Frontier is still basically untouched and very broken. The only way out of it is to Fly back to Johto, as the boat doesn't work correctly. The Pike, Pyramid, and Factory have missing maps still. Day/Night palettes are also incorrectly applied on the Frontier tileset.