Refactor Apricorn Ball Code
Sierra's Apricorn Ball implementation did not play nice with the GCN games and showed DPPt balls when migrated up to Gen 4.

This new implementation uses padding in the Pokemon struct to store Apricorn Balls. Like with HGSS Apricorn Balls, other games will see Pokemon caught in them as being in Poke Balls which fixes the issues with the GCN and Gen 4 games.