Do GS Ball Event
The GS Ball event and battling Celebi is completely done. Nothing enables it yet, but setting `FLAG_GS_BALL_AVAILABLE_FROM_POKECOM_CENTER` is all that is required to do so.

Celebi also regenerates if not caught after entering the Hall of Fame, with new text at the shrine because the player already placed the GS Ball there when they failed to catch it the first time. For rematches, they "disturb" the GS Ball that is already there.

Additionally, refactored Kurt and his granddaughter's scripts, as there were several inaccuracies to Crystal. Kurt's granddaughter now speaks about her dad once the player gets to Vermilion, which was missed when I was scripting Kanto.

Additionally, the Slowpoke in Azalea Town now all start with their tails chopped off, but once the player gets to Vermilion City they all grow back.