Script Start of Sevii Quest
The player now receives a phone call from Bill after unlocking Mt. Silver telling them to meet him in Cinnabar to come with him to Sevii. Everything up to leaving Kanto is scripted, but the script dumps you back to Cinnabar because the Sevii Islands aren't in yet.

Inserted the Seagallop OW for this event.