Time Capsule Event Safeguard
In GSC, if the player happens to never enter Ecruteak's Pokemon Center, they will never meet Bill nor unlock the Time Capsule.

This is an issue for CrystalDust, because Bill invites the player to the Sevii Islands and so needs to actually know them.

Now, for the rare player that never enters the Ecruteak Pokemon Center, Bill's unlocking the Time Capsule event will happen in the Indigo Plateau Pokemon Center, as that Pokemon Center is mandatory to enter.

The original event still happens in Ecruteak, which will disable this safeguard if triggered.