Update Youngster Joey Rematch Text
Fixes weird placement in the textbox.

I have a suspicion that none of the script flow for getting a trainer's number is accurate to Crystal, but don't have the time to check it right now.

Additionally, ALL text strings from this PR need to be checked for textbox overflow and placement. Youngster Joey overflows the Phone textbox with "one-and-only", for example.