Avert Potential Softlock in Indigo Plateau
There was a potential softlock on the Route 26/27/Victory Road/Indigo Plateau circuit after becoming Champion but before beating Snorlax, since that requires Surf or Fly to get out if you can't beat the Elite Four with what you have.

If the player is the Champ but the Snorlax hasn't been defeated to open up the badge check gate path to Viridian City, the game checks for the ability to use/learn Surf, as well as the ability to use/learn Fly to get to another Kanto location (if the player hasn't visited Vermilion yet, they will only be able to fly to Indigo Plateau so they'd still be stuck). If the player can't do either of those things, the Abra man respawns and can teleport the player home.

Used ChatGPT to optimize both the Cianwood/Cinnabar and the Indigo Plateau softlock-checking functions, since I was getting lost in the sauce with all the checks. After thorough testing, what it wrote works perfectly and is much easier to read.