Start Moving NPC Phone Calls to Scripts
Refactored a lot of the phone stuff so that it could be moved to generic scripts.

Youngster Joey is now entirely accurate to Crystal. No other trainers have the new script handling, and because the GenericText table was clobbered, will probably crash the game if they call.

Trainers that have the same phone script flow as Joey can be set up with essentially a two-line script, as Joey's script has been genericized. Hopefully, trainers that just rematch and don't gossip, activate events, or give items actually are the same as Joey's script.

Non-forced rematches now use the flags in `gSaveBlock1Ptr->trainerRematches` instead of a regular flag. Forced rematch flags have been reduced to only 4, as that's the maximum amount of trainers that you can force to rematch in a single day. This reclaimed 32 daily flags. Queued rematches are cleared daily now.

Note to self, figure out what to do with the singular generic texts. I suspect that each trainer actually only has one set and the extras are for gossiping. Can probably remove the table altogether.