Optimize Several script.c Functions
The main takeaway from this commit is that there should be significantly less lag while the Tin Tower/Eusine checks whether or not you own all the legendary beasts under typical circumstances, as the functions that do the checking are much more efficient. The worst-case scenario is a box and party full of legendary beasts that match the player's OTID but not their name.

Additionally, the setup functions for big and small dolls in the player's room now use lookup tables instead of giant switch statements because using switch statements in the first place was stupid.