4/28 Phone Trainers Done; Fix Rematch-Related Bugs
With this commit, Reena, Gaven, and Beth are accurate to Crystal. All their text has been checked for textbox placement and overflow. Joey was done in a previous commit.

`gSaveBlock1Ptr->trainerRematches` is now treated as bitflags for simplicity and for potential space savings in the saveblock. The size of this field hasn't been reduced in the saveblock yet, but with the vanilla trainers could get down to only 6 bytes (padded to 8).

Rematch states were not getting checked or cleared properly, but this commit fixes that.

Also removed or commented out some irrelevant/unused functions or bits of functions that used the old `gSaveBlock1Ptr->trainerRematches` way of doing things.